feat(hitl): move core_api from provider to core
